refactor: 패키지 이동 - PostEditorResponse 패키지 request에서 response로 이동
