+++ android/app/src/main/java/com/app/edonymyeon/data/dto/response/PostEditorResponse.kt
@@ -1,4 +1,4 @@
-package com.app.edonymyeon.data.dto.request
+package com.app.edonymyeon.data.dto.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
